requirements.txt
openai
langchain
langchain-community
langchain-core
tiktoken
faiss-cpu
python-dotenv
markdown
fastapi
uvicorn